# samples/client/petstore/go/go-petstore/api/openapi.yaml
openapi: 3.0.1
info:
  description: "This spec is mainly for testing Petstore server and contains fake\
    \ endpoints, models. Please do not use this for any other purpose. Special characters:\
    \ \" \\"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAPI Petstore
  version: 1.0.0
servers:
- url: http://petstore.swagger.io:80/v2
...
paths:
  /pet:
    post:
      operationId: addPet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
          application/xml:
            schema:
              $ref: '#/components/schemas/Pet'
        description: Pet object that needs to be added to the store
        required: true
      responses:
        "200":
          content: {}
          description: successful operation
        "405":
          content: {}
          description: Invalid input
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Add a new pet to the store
      tags:
      - pet
      x-codegen-request-body-name: body
    put:
      operationId: updatePet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
          application/xml:
            schema:
              $ref: '#/components/schemas/Pet'
        description: Pet object that needs to be added to the store
        required: true
      responses:
        "200":
          content: {}
          description: successful operation
        "400":
          content: {}
          description: Invalid ID supplied
        "404":
          content: {}
          description: Pet not found
        "405":
          content: {}
          description: Validation exception
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Update an existing pet
      tags:
      - pet
      x-codegen-request-body-name: body
  /pet/findByStatus:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: findPetsByStatus
      parameters:
      - description: Status values that need to be considered for filter
        explode: false
        in: query
        name: status
        required: true
        schema:
          items:
            default: available
            enum:
            - available
            - pending
            - sold
            type: string
          type: array
        style: form
      - in: query
        name: scalarInt32
        schema:
          format: int32
          type: integer
      - in: query
        name: scalarInt64
        schema:
          format: int64
          type: integer
      - in: query
        name: scalarFloat32
        schema:
          format: float
          type: number
      - in: query
        name: scalarFloat64
        schema:
          format: double
          type: number
      - explode: false
        in: query
        name: arrayInt32
        schema:
          items:
            format: int32
            type: integer
          type: array
        style: form
      - explode: false
        in: query
        name: arrayInt64
        schema:
          items:
            format: int64
            type: integer
          type: array
        style: form
      - explode: false
        in: query
        name: arrayFloat32
        schema:
          items:
            format: float
            type: number
          type: array
        style: form
      - explode: false
        in: query
        name: arrayFloat64
        schema:
          items:
            format: double
            type: number
          type: array
        style: form
      - in: query
        name: scalarString
        schema:
          type: string
      - in: query
        name: scalarBoolean
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Pet'
                type: array
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Pet'
                type: array
          description: successful operation
        "400":
          content: {}
          description: Invalid status value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Finds Pets by status
      tags:
      - pet
  /pet/findByTags:
    get:
      deprecated: true
      description: "Multiple tags can be provided with comma separated strings. Use\
        \ tag1, tag2, tag3 for testing."
      operationId: findPetsByTags
      parameters:
      - description: Tags to filter by
        explode: false
        in: query
        name: tags
        required: true
        schema:
          items:
            type: string
          type: array
          uniqueItems: true
        style: form
      responses:
        "200":
          content:
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Pet'
                type: array
                uniqueItems: true
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Pet'
                type: array
                uniqueItems: true
          description: successful operation
        "400":
          content: {}
          description: Invalid tag value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Finds Pets by tags
      tags:
      - pet
  /pet/{petId}/{scalarInt32}/{scalarInt64}/{scalarFloat32}/{scalarFloat64}:
    delete:
      operationId: deletePet
      parameters:
      - in: header
        name: api_key
        schema:
          type: string
      - description: Pet id to delete
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: scalarInt32
        required: true
        schema:
          format: int32
          type: integer
      - in: path
        name: scalarInt64
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: scalarFloat32
        required: true
        schema:
          format: float
          type: number
      - in: path
        name: scalarFloat64
        required: true
        schema:
          format: double
          type: number
      responses:
        "200":
          content: {}
          description: successful operation
        "400":
          content: {}
          description: Invalid pet value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Deletes a pet
      tags:
      - pet
    get:
      description: Returns a single pet
      operationId: getPetById
      parameters:
      - description: ID of pet to return
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: scalarInt32
        required: true
        schema:
          format: int32
          type: integer
      - in: path
        name: scalarInt64
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: scalarFloat32
        required: true
        schema:
          format: float
          type: number
      - in: path
        name: scalarFloat64
        required: true
        schema:
          format: double
          type: number
      responses:
        "200":
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Pet'
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
          description: successful operation
        "400":
          content: {}
          description: Invalid ID supplied
        "404":
          content: {}
          description: Pet not found
      security:
      - api_key: []
      summary: Find pet by ID
      tags:
      - pet
    post:
      operationId: updatePetWithForm
      parameters:
      - description: ID of pet that needs to be updated
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: scalarInt32
        required: true
        schema:
          format: int32
          type: integer
      - in: path
        name: scalarInt64
        required: true
        schema:
          format: int64
          type: integer
      - in: path
        name: scalarFloat32
        required: true
        schema:
          format: float
          type: number
      - in: path
        name: scalarFloat64
        required: true
        schema:
          format: double
          type: number
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/updatePetWithForm_request'
      responses:
        "405":
          content: {}
          description: Invalid input
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Updates a pet in the store with form data
      tags:
      - pet
  /pet/{petId}/uploadImage:
    post:
      operationId: uploadFile
      parameters:
      - description: ID of pet to update
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/uploadFile_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: uploads an image
      tags:
      - pet
...
components:
  schemas:
...
    Category:
      example:
        name: default-name
        id: 6
      properties:
        id:
          format: int64
          type: integer
        name:
          default: default-name
          type: string
      required:
      - name
      type: object
      xml:
        name: Category
...
    Tag:
      example:
        name: name
        id: 1
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
      type: object
      xml:
        name: Tag
    Pet:
      example:
        scalar_float64: 35.00003813479535
        photoUrls:
        - photoUrls
        - photoUrls
        scalar_boolean: true
        scalar_float32: 22.982893
        tags:
        - name: name
          id: 1
        - name: name
          id: 1
        scalar_string: "19"
        array_float32:
        - 22.288486
        - 22.288486
        array_float64:
        - 27.637660275456938
        - 27.637660275456938
        array_int64:
        - 26
        - 26
        name: doggie
        array_int32:
        - 40
        - 40
        scalar_int32: 31
        id: 0
        scalar_int64: 31
        category:
          name: default-name
          id: 6
        status: available
      properties:
        id:
          format: int64
          type: integer
          x-is-unique: true
        category:
          $ref: '#/components/schemas/Category'
        name:
          example: doggie
          type: string
        photoUrls:
          items:
            type: string
          type: array
          uniqueItems: true
          xml:
            name: photoUrl
            wrapped: true
        tags:
          items:
            $ref: '#/components/schemas/Tag'
          type: array
          xml:
            name: tag
            wrapped: true
        status:
          description: pet status in the store
          enum:
          - available
          - pending
          - sold
          type: string
        scalar_int32:
          default: 19
          format: int32
          maximum: 42
          minimum: 17
          type: integer
        scalar_int64:
          default: 19
          format: int64
          maximum: 42
          minimum: 17
          type: integer
        scalar_float32:
          default: 19.19
          format: float
          maximum: 42.42
          minimum: 17.17
          type: number
        scalar_float64:
          default: 19.19
          format: double
          maximum: 42.42
          minimum: 17.17
          type: number
        array_int32:
          items:
            format: int32
            maximum: 42
            minimum: 17
            type: integer
          type: array
        array_int64:
          items:
            format: int64
            maximum: 42
            minimum: 17
            type: integer
          type: array
        array_float32:
          items:
            format: float
            maximum: 42.42
            minimum: 17.17
            type: number
          type: array
        array_float64:
          items:
            format: double
            maximum: 42.42
            minimum: 17.17
            type: number
          type: array
        scalar_string:
          default: "19"
          type: string
        scalar_boolean:
          default: true
          type: boolean
      required:
      - name
      - photoUrls
      type: object
      xml:
        name: Pet
    ApiResponse:
      example:
        code: 0
        type: type
        message: message
      properties:
        code:
          format: int32
          type: integer
        type:
          type: string
        message:
          type: string
      type: object
...
    updatePetWithForm_request:
      properties:
        name:
          description: Updated name of the pet
          type: string
        status:
          description: Updated status of the pet
          type: string
      type: object
    uploadFile_request:
      properties:
        additionalMetadata:
          description: Additional data to pass to server
          type: string
        file:
          description: file to upload
          format: binary
          type: string
      type: object
...
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/api/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey
    api_key_query:
      in: query
      name: api_key_query
      type: apiKey
    http_basic_test:
      scheme: basic
      type: http
...